Adding Security in Journal Controllers

+++ journal-app/src/main/java/com/mehedy/journal_app/service/JournalEntryService.java
@@ -6,7 +6,6 @@ import com.mehedy.journal_app.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -17,15 +16,15 @@ public class JournalEntryService {
     private final JournalEntryRepository journalEntryRepository;
     private final UserService userService;
 
-    public JournalEntryService(JournalEntryRepository journalEntryRepository,  UserService userService) {
+    public JournalEntryService(JournalEntryRepository journalEntryRepository, UserService userService) {
         this.journalEntryRepository = journalEntryRepository;
-        this.userService=userService;
+        this.userService = userService;
     }
 
     public void saveEntry(JournalEntry journalEntry, String username) {
-        User user=userService.findByUserName(username);
+        User user = userService.findByUserName(username);
         journalEntry.setDate(LocalDateTime.now());
-        JournalEntry saved= journalEntryRepository.save(journalEntry);
+        JournalEntry saved = journalEntryRepository.save(journalEntry);
         user.getJournalEntries().add(saved);
         userService.save(user);
     }
@@ -36,24 +35,36 @@ public class JournalEntryService {
     }
 
 
-
-    public List<JournalEntry> getAll(){
+    public List<JournalEntry> getAll() {
         return journalEntryRepository.findAll();
     }
 
-    public Optional<JournalEntry> findById(ObjectId id)
-    {
+    public Optional<JournalEntry> findById(ObjectId id) {
         return journalEntryRepository.findById(id);
     }
 
 
-    public void deleteById(ObjectId id,String username)
-    {
-        User user=userService.findByUserName(username);
-        // In no relational db there is no option of cascade, so we delete the data manually.
-        user.getJournalEntries().removeIf(journalEntry -> journalEntry.getId().equals(id));
-        userService.save(user);
-        journalEntryRepository.deleteById(id);
+    //    @Transactional
+    public boolean deleteById(ObjectId id, String username) {
+        boolean removed = false;
+
+        try {
+            User user = userService.findByUserName(username);
+            // In no relational db there is no option of cascade, so we delete the data manually.
+            removed = user.getJournalEntries().removeIf(journalEntry -> journalEntry.getId().equals(id));
+            if (removed) {
+                userService.save(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("An error occurred while trying to delete the journal entry", e);
+        }
+
+        return removed;
+    }
+
+    public List<JournalEntry> findByUsername(String username) {
+        return null;
     }
 
 }
